Fix body click animation to only work on .body
The animation previously triggered when clicking any child div of .body, and the workaround of making the limbs yellow to match the body change was annoying. I fixed this by adding a child div .body-click that is a transparent layer on top of just the body, and is the div that triggers the event and hosts the spongebob background image.