Fix head click
Previously, hat event would also trigger (with wink event) when eyes were clicked. Fixed this by creating a head-click child of head, and hat-top and hat-brim children of head-click. The head-click div is set behind the eyes in the same position as the head itself, and the hat-top and hat-brim divs are invisible until the head-click div is clicked.

Now all the (current) animations do what I want them to do! Yay!